Fixed compilation error under Windows

+++ platforms/reference/tests/TestReferenceMonteCarloBarostat.cpp
@@ -117,7 +117,7 @@ void testIdealGas() {
         }
         volume /= steps;
         double expected = numParticles*BOLTZ*temp[i]/pressureInMD;//+numParticles*(4*M_PI/3)*sigma*sigma*sigma;
-        ASSERT_USUALLY_EQUAL_TOL(expected, volume, 3/std::sqrt(steps));
+        ASSERT_USUALLY_EQUAL_TOL(expected, volume, 3/std::sqrt((double) steps));
     }
 }
